code refactoring appJs to separate components
