+++ src/App.js
@@ -3,10 +3,16 @@ import Header from './components/dieComponents/Header';
 import NewDie from './components/dieComponents/Newdie';
 import Die from './components/dieComponents/Die';
 import { nanoid } from 'nanoid';
+import { useEffect } from 'react';
 
 function App() {
 
-  const[dice, setDice] = React.useState(allNewDice())
+
+  const [tenzies, setTenzies] = React.useState(false);
+
+  useEffect(() => {
+    console.log('Dice state changed')
+  }, [dice])
 
   function generateNewDie() {
     return {
@@ -16,30 +22,22 @@ function App() {
     }
   }
 
-  function allNewDie() {
-    const newDice = []
-    for (let i = 0; i < 10; i++){
-      newDice.push(generateNewDice())
-    }
-    return newDice
-    }
+    // function holdDice(id) {
+    //   setDice(oldDice => oldDiceDice.map(die =>{
+    //     return die.id === id ?
+    //     {...die, isHeld: !die.isHeld}:
+    //     die
+    //   }))
+    // }
 
-    function holdDice(id) {
-      setDice(oldDice => oldDiceDice.map(die =>{
-        return die.id === id ?
-        {...die, isHeld: !die.isHeld}:
-        die
-      }))
-    }
-
-    function diceElements = dice.map(die => (
-      <Die
-        key={die.id}
-        value ={value.id}
-        isHeld={die.isHeld}
-        holdDice={() => holdDice(die.id)}
-      />
-    ))
+    // function diceElements = dice.map(die => (
+    //   <Die
+    //     key={die.id}
+    //     value ={value.id}
+    //     isHeld={die.isHeld}
+    //     holdDice={() => holdDice(die.id)}
+    //   />
+    // ))
 
   return (
     <div className="App">
@@ -47,7 +45,9 @@ function App() {
       <div className= 'dice-container'>
         {diceElements}
       </div>
+      <NewDie/>
       <button className='roll--btn' onClick={rollDice}>Roll</button>
+
     </div>
   );
 }
