+++ src/components/dieComponents/Newdie.js
@@ -4,9 +4,13 @@ import React from 'react';
 
 export default function Newdie() {
 
+  const newDice = []
+    for (let i = 0; i < 10; i++){
+      newDice.push(generateNewDice())
+    }
   return (
     <div>
-
+      {newDice}
     </div>
     // <Container>
     //   <Button variant="primary">Roll</Button>
